+++ contracts/multisig-improved/src/common_types/mod.rs
@@ -1,0 +1,2 @@
+pub mod action;
+pub mod user_role;
