+++ contracts/multisig-improved/src/common_types/user_role.rs
