+++ contracts/multisig-improved/wasm-multisig-full/src/lib.rs
@@ -53,14 +53,14 @@ multiversx_sc_wasm_adapter::endpoints! {
         performAction => perform_action_endpoint
         performBatch => perform_batch
         dnsRegister => dns_register
-        getPendingActionFullInfo => get_pending_action_full_info
-        userRole => user_role
-        getAllBoardMembers => get_all_board_members
-        getAllProposers => get_all_proposers
         getActionData => get_action_data
         getActionSigners => get_action_signers
         getActionSignerCount => get_action_signer_count
         getActionValidSignerCount => get_action_valid_signer_count
+        getPendingActionFullInfo => get_pending_action_full_info
+        userRole => user_role
+        getAllBoardMembers => get_all_board_members
+        getAllProposers => get_all_proposers
     )
 }
